Issue 2892(part2) - AgentAutoUpgrade: agent auto upgrade for edge cluster
